project: isolate cli
